requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pypdf
python-multipart